functions/azureShutdownNotify/requirements.txt
pushbullet.py==0.10.0
bitly_api==0.3
requests
